Tentative logic.
Basically it checks a list of all current appointments. IF the new start time is less than or equal to the end time of a current one AND the end time of the same existing appointment is less than or equal to the new end time then the appointment isn't made.